refactor: corrigi o import do Timestamp e altera os id's para Long
